rebase on alben's newest commits, remove duplicate header, migrate code to LowPower.cpp.
